requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Vector databases
chromadb>=0.4.14
pinecone-client>=3.0.0

# Embeddings and LLMs
openai>=1.10.0
langchain>=0.1.0
langchain-openai>=0.0.5
sentence-transformers>=2.2.2

# PDF and document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
pdfplumber>=0.10.3

# ML and ML monitoring
scikit-learn>=1.3.2
xgboost>=2.0.2
mlflow>=2.9.1
evidently>=0.4.0

# Monitoring and metrics
prometheus-client>=0.19.0

# Security and authentication
PyJWT>=2.8.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

# Development and debugging
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.2
pre-commit>=3.5.0

# Logging
python-json-logger>=2.0.7

# Utilities
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.3

# API documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.0
